third_party/WebKit/LayoutTests/inspector-protocol/css/css-get-rule-list.html

<html>
<head>
<script type="text/javascript" src="resources/inspector-protocol-test.js"></script>
<script type="text/javascript" src="resources/css-protocol-test.js"></script>
<script type="text/javascript" src="resources/dom-protocol-test.js"></script>
<script type="text/javascript">
function test()
{
    InspectorTest.sendCommandOrDie("DOM.enable", {});
    InspectorTest.sendCommandOrDie("CSS.enable", {}, () => testGetRuleListAfterStartingRecording());

    function testGetRuleListAfterStartingRecording()
    {
        InspectorTest.sendCommandOrDie("CSS.startRuleUsageTracking", {},  () =>
                                       { InspectorTest.sendCommandOrDie("CSS.stopRuleUsageTracking", {}, ruleListLoaded); });
    }

    function ruleListLoaded(result)
    {
        if (!result) {
            InspectorTest.completeTest();
            return;
        }
        var rules = result.ruleUsage;
        rules.sort((a, b) => a.startOffset - b.startOffset);
        var usedLines = rules.filter(rule => rule.used);
        var unusedLines = rules.filter(rule => !rule.used);

        InspectorTest.log("Used rules offsets: " + usedLines.length);
        for (var line of usedLines)
            InspectorTest.log(line.startOffset);

        InspectorTest.log("Unused rules offsets: " + unusedLines.length);
        for (var line of unusedLines)
            InspectorTest.log(line.startOffset);

        InspectorTest.completeTest();
    }
}

</script>
<link rel="stylesheet" href="resources/coverage.css">
</head>
<body onload="runTest();">
<h1 class="class">Class Selector</h1>
<p id="id">ID Selector</p>
<div></div>

</body>
</html>
